+++ client/package.json
@@ -2,6 +2,7 @@
   "name": "glassfiller-react",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:3001/",
   "dependencies": {
     "@apollo/client": "^3.6.9",
     "@fortawesome/free-solid-svg-icons": "^6.1.2",
