+++ client/src/App.js
@@ -4,6 +4,7 @@ import { Routes, Route } from 'react-router-dom';
 import MyBar from './pages/MyBar';
 import Home from './pages/Home.js';
 import Login from './pages/Login';
+import Signup from './pages/Signup.js';
 import NavBar from './components/NavBar';
 import Footer from './components/Footer';
 import { setContext } from '@apollo/client/link/context';
@@ -37,18 +38,19 @@ const client = new ApolloClient({
 function App() {
   return (
     <ApolloProvider client={client}>
-  <div> 
-    <NavBar />
-   
-    <Routes>
-      <Route exact path="/" element={<Home />} />
-      <Route exact path="/mybar" element={<MyBar />}/>
-      <Route exact path="/login" element={<Login />}/>
-    </Routes>
-    
-    <Footer />
-    
-    </div>
+      <div>
+        <NavBar />
+
+        <Routes>
+          <Route exact path="/" element={<Home />} />
+          <Route exact path="/mybar" element={<MyBar />} />
+          <Route exact path="/login" element={<Login />} />
+          <Route exact path="/signup" element={<Signup />} />
+        </Routes>
+
+        <Footer />
+
+      </div>
     </ApolloProvider>
   )
 }
