--- client/src/components/Login/index.js
@@ -1,81 +1,0 @@
-import React, { useState } from 'react';
-
-import { useMutation } from '@apollo/client';
-import { LOGIN_USER } from '../../utils/mutations';
-
-
-import Auth from '../../utils/auth';
-
-const Login = (props) => {
-    const [formState, setFormState] = useState({ email: '', password: '' });
-    const [login, { error }] = useMutation(LOGIN_USER);
-
-    const handleChange = (event) => {
-        const { name, value } = event.target;
-    
-        setFormState({
-          ...formState,
-          [name]: value,
-        });
-      };
-    
-      // submit form
-      const handleFormSubmit = async (event) => {
-        event.preventDefault();
-    
-        try {
-          const { data } = await login({
-            variables: { ...formState },
-          });
-    
-          Auth.login(data.login.token);
-        } catch (e) {
-          console.error(e);
-        }
-    
-        // clear form values
-        setFormState({
-          email: '',
-          password: '',
-        });
-      };
-
-return (
-    <main>
-        <div>
-            <div className='card'>
-                <h3 className='card-header'>Login</h3>
-                <div className='card-body'>
-                <form onSubmit={handleFormSubmit}>
-                <input
-                className="form-input"
-                placeholder="Your email"
-                name="email"
-                type="email"
-                id="email"
-                value={formState.email}
-                onChange={handleChange}
-              />
-              <input
-                className="form-input"
-                placeholder="******"
-                name="password"
-                type="password"
-                id="password"
-                value={formState.password}
-                onChange={handleChange}
-              />
-              <button type='submit'>
-                Submit
-              </button>
-              </form>
-
-              {error && <div>Login failed</div>}
-                </div>
-            </div>
-        </div>
-    </main>
-);
-};
-
-export default Login;
