+++ client/src/utils/auth.js
@@ -1,47 +1,47 @@
-// import decode from 'jwt-decode';
+import decode from 'jwt-decode';
 
 
-// class AuthService {
-//   getProfile() {
-//     return decode(this.getToken());
-//   }
+class AuthService {
+    getProfile() {
+        return decode(this.getToken());
+    }
 
-//   loggedIn() {
-//     // Checks if there is a saved token and it's still valid
-//     const token = this.getToken();
-//     return !!token && !this.isTokenExpired(token); 
-//   }
+    loggedIn() {
+        // Checks if there is a saved token and it's still valid
+        const token = this.getToken();
+        return !!token && !this.isTokenExpired(token);
+    }
 
-//   isTokenExpired(token) {
-//     try {
-//       const decoded = decode(token);
-//       if (decoded.exp < Date.now() / 1000) {
-//         return true;
-//       } else return false;
-//     } catch (err) {
-//       return false;
-//     }
-//   }
+    isTokenExpired(token) {
+        try {
+            const decoded = decode(token);
+            if (decoded.exp < Date.now() / 1000) {
+                return true;
+            } else return false;
+        } catch (err) {
+            return false;
+        }
+    }
 
-//   getToken() {
-   
-//     return localStorage.getItem('id_token');
-//   }
+    getToken() {
 
-//   login(idToken) {
-//     // Saves user token to localStorage
-//     localStorage.setItem('id_token', idToken);
+        return localStorage.getItem('id_token');
+    }
 
-//     window.location.assign('/');
-//   }
+    login(idToken) {
+        // Saves user token to localStorage
+        localStorage.setItem('id_token', idToken);
 
-//   logout() {
-//     // Clear user token and profile data from localStorage
-//     // axios.defaults.headers.common["Authorization"] = null;
-//     localStorage.removeItem('id_token');
-//     // this will reload the page and reset the state of the application
-//     window.location.assign('/');
-//   }
-// }
+        window.location.assign('/');
+    }
 
-// export default new AuthService();
+    logout() {
+        // Clear user token and profile data from localStorage
+        // axios.defaults.headers.common["Authorization"] = null;
+        localStorage.removeItem('id_token');
+        // this will reload the page and reset the state of the application
+        window.location.assign('/');
+    }
+}
+
+export default new AuthService();
