+++ server/schemas/typeDefs.js
@@ -5,6 +5,7 @@ const typeDefs = gql`
     _id: ID
     username: String
     email: String
+    password: String
     recipes: [Recipe]
   }
